Move coordinator for TwenteMilieu into own module
