+++ homeassistant/components/twentemilieu/__init__.py
@@ -2,53 +2,25 @@
 
 from __future__ import annotations
 
-from datetime import date, timedelta
-
-from twentemilieu import TwenteMilieu, WasteType
 import voluptuous as vol
 
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_ID, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import config_validation as cv
-from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
-from .const import CONF_HOUSE_LETTER, CONF_HOUSE_NUMBER, CONF_POST_CODE, DOMAIN, LOGGER
-
-SCAN_INTERVAL = timedelta(seconds=3600)
+from .coordinator import TwenteMilieuConfigEntry, TwenteMilieuDataUpdateCoordinator
 
 SERVICE_UPDATE = "update"
 SERVICE_SCHEMA = vol.Schema({vol.Optional(CONF_ID): cv.string})
 
 PLATFORMS = [Platform.CALENDAR, Platform.SENSOR]
 
-type TwenteMilieuDataUpdateCoordinator = DataUpdateCoordinator[
-    dict[WasteType, list[date]]
-]
-type TwenteMilieuConfigEntry = ConfigEntry[TwenteMilieuDataUpdateCoordinator]
-
 
 async def async_setup_entry(
     hass: HomeAssistant, entry: TwenteMilieuConfigEntry
 ) -> bool:
     """Set up Twente Milieu from a config entry."""
-    session = async_get_clientsession(hass)
-    twentemilieu = TwenteMilieu(
-        post_code=entry.data[CONF_POST_CODE],
-        house_number=entry.data[CONF_HOUSE_NUMBER],
-        house_letter=entry.data[CONF_HOUSE_LETTER],
-        session=session,
-    )
-
-    coordinator: TwenteMilieuDataUpdateCoordinator = DataUpdateCoordinator(
-        hass,
-        LOGGER,
-        config_entry=entry,
-        name=DOMAIN,
-        update_interval=SCAN_INTERVAL,
-        update_method=twentemilieu.update,
-    )
+    coordinator = TwenteMilieuDataUpdateCoordinator(hass, entry)
     await coordinator.async_config_entry_first_refresh()
 
     entry.runtime_data = coordinator
